pyproject.toml
[project]
name = "aim-hot-mess-coach"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.2",
    "uvicorn>=0.38.0",
    "python-multipart>=0.0.9",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "opencv-python>=4.8.0",
]
